// libraries/element-pro/design/select/type.ts
export interface ElOptionProps {
  /**
   * 当前选项是否为全选，全选可以在顶部，也可以在底部。点击当前选项会选中禁用态除外的全部选项，即使是分组选择器也会选中全部选项
   * @default false
   */
  checkAll?: boolean;
  /**
   * 用于定义复杂的选项内容
   */
  content?: string | ElNode;
  /**
   * 用于定义复杂的选项内容。同 content
   */
  default?: string | ElNode;
  /**
   * 是否禁用该选项
   * @default false
   */
  disabled?: boolean;
  /**
   * 选项名称
   * @default ''
   */
  label?: string;
  /**
   * 选项标题，在选项过长时hover选项展示
   * @default ''
   */
  title?: string;
  /**
   * 选项值
   */
  value?: string | number | boolean;
}

export interface ElOptionGroupProps {
  /**
   * 是否显示分隔线
   * @default true
   */
  divider?: boolean;
  /**
   * 分组别名
   * @default ''
   */
  label?: string;
}

export type SelectValue<T extends SelectOption = SelectOption> = string | number | boolean | T | Array<SelectValue<T>>;

export type SelectValueChangeTrigger = 'clear' | 'tag-remove' | 'backspace' | 'check' | 'uncheck' | 'default';

export interface SelectRemoveContext<T> {
  value: string | number;
  data: T;
  e: MouseEvent | KeyboardEvent;
}

export type SelectOption = ElOptionProps | SelectOptionGroup | PlainObject;

export interface SelectOptionGroup extends ElOptionGroupProps {
  group: string;
  children: Array<ElOptionProps>;
}